Add MyAdvice class contains @ExceptionHandler method
