+++ src/main/java/com/codessquad/qna/MyAdvice.java
@@ -1,0 +1,19 @@
+package com.codessquad.qna;
+
+import javassist.NotFoundException;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+public class MyAdvice {
+
+    @ExceptionHandler(NotFoundException.class)
+    public String handleNotFountException() {
+        return "notExistHandle";
+    }
+
+    @ExceptionHandler(IllegalStateException.class)
+    public String handleIllegalStateException() {
+        return "unableToAccessToOthers";
+    }
+}
